pyproject.toml
[tool.poetry]
name = "drfwn-quick"
version = "0.1.0"
description = ""
authors = ["LamerLink <36551116+LamerLink@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-writable-nested = "^0.7.0"

[tool.poetry.group.dev.dependencies]
bump2version = "^1.0.1"
flake8 = "^6.0.0"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
